[CAS-335] Use a single Call implementation in tests
